Performance work -- attempt to reduce task based locking.
We have tried to use a lock free algorithm for the individual tasks.
This seems to lead to a reduced need for locking, but higher level code
still needs locks.

This code works, but the latency figures are inexplicably higher
than they were before.  Not sure why yet -- shorter code paths
may leading to more contention on shared locks?  Or changes in
task dispatching?  Or cache aliasing or something similar?